Changed datatype to 64 bit integer

+++ src/Entity/User.php
@@ -91,12 +91,12 @@ class User
         return $this;
     }
 
-    public function getChatId(): ?float
+    public function getChatId(): ?int
     {
         return $this->chat_id;
     }
 
-    public function setChatId(float $chat_id): self
+    public function setChatId(int $chat_id): self
     {
         $this->chat_id = $chat_id;
